Allow static configuration of verified desktops
This is useful when configured to only start sessions on remote hosts.
In that case, it is better to allow the verified desktops to be
set in the configuration file than to try to determine them from the
verfied desktops on localhost.

An alternative, might be to have the verfied desktops determined from
one of the remote hosts.  But that raises questions such as which host;
which user; if the `root` user, is passwordless SSH access going to be
available.