Don't pre-verify desktops
For some desktop types such as `gnome`, the `verify` command may only be
successful when ran by the root user.  We don't want to force pre-verify
the desktop as the correct global-verification state will be overwritten
with a possibly wrong user-verification state.

We stil want to attempt to verify the desktop state if it fails to
launch due to it not being verified.